% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WtCoxG.R
\name{fun.optimalWeight}
\alias{fun.optimalWeight}
\title{optimal weight for mu_ext -----------------------------------------------}
\usage{
fun.optimalWeight(par, pop.prev, R, y, mu1, w, mu, N, n.ext, sigma2, TPR)
}
\description{
optimal weight for mu_ext -----------------------------------------------
}
