% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WtCoxG.R
\name{QCforBatchEffect}
\alias{QCforBatchEffect}
\title{Quality control to check batch effect between study cohort and reference population.}
\usage{
QCforBatchEffect(
  GenoFile = NULL,
  GenoFileIndex = NULL,
  OutputFile,
  control = list(AlleleOrder = "ref-first"),
  PhenoData,
  RefAfFile,
  RefPrevalence,
  SNPnum = 10000,
  sparseGRM = NULL
)
}
\arguments{
\item{GenoFile}{A character string of the genotype file. See Details section for more details.}

\item{GenoFileIndex}{Additional index file(s) corresponding to GenoFile. See Details section for more details.}

\item{control}{A list of parameters to decide which markers to extract. See \code{Details} section for more details.}

\item{PhenoData}{A dataframe that must have at least two columns: \code{SampleID} holds the personal identifiers for all individuals, and \code{Indicator} holds whether the event occurred (0 or 1 or NA).}

\item{SNPnum}{An integer specifying the minimum number of markers. The default is 1e4.}

\item{sparseGRM}{a three-column sparse GRM file with the first column as "ID1",the second column as "ID2", and the last column as "Value" (i.e., two times of kinship coefficient) without information of distant genetic relatedness (such as population structure).}

\item{RefAFfile}{A character string of the reference file. The reference file must be a \code{txt} file (header required) including at least 7 columns: \code{CHROM}, \code{POS}, \code{ID}, \code{REF}, \code{ALT}, \code{AF_ref}, \code{AN_ref}.}

\item{RefPrev}{A numeric value of the event rate in the population.}
}
\value{
An R object with a class of "QCforBatchEffect".
\itemize{
\item{mergeGenoInfo}: A dataframe of marker info and reference MAF.
\item{cutoff}: A numeric, the cut-off of batch effect.
\item{count}: A dataframe of the frequency of the batch effect p-value.
\item{PhenoData}: A dataframe of the input PhenoData.
\item{control}: A list of parameters to decide which markers to extract.
}
}
\description{
This function performs quality control to test for the batch effect between a study cohort and a reference population. And fit a weighted null model.
}
\examples{
setwd(system.file("", package = "WtCoxG"))
PhenoData = read.table("simuPHENO_WtSPAG.txt", header = T)
RefPrevalence = 0.1

obj.WtCoxG = QCforBatchEffect(GenoFile = "simuBGEN1.bgen",
                             GenoFileIndex = c("simuBGEN1.bgen.bgi",
                                                "simuBGEN1.sample"),
                             OutputFile = "qcBGEN1.txt",
                             control=list(AlleleOrder = "ref-first",
                                          AllMarkers = T,
                                          IndicatorColumn = "SurvEvent", SampleIDColumn = "IID"),
                             PhenoData=PhenoData,
                             RefAfFile = "RefMAFs.txt",
                             RefPrevalence = RefPrevalence,
                             SNPnum=1e4)
names(obj.WtCoxG)
}
