% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WtCoxG.R
\name{TestforBatchEffect}
\alias{TestforBatchEffect}
\title{Quality control to check batch effect between study cohort and reference population.}
\usage{
TestforBatchEffect(
  GenoFile = NULL,
  GenoFileIndex = NULL,
  Geno.mtx = NULL,
  OutputFile,
  control = list(AlleleOrder = "ref-first"),
  PhenoFile,
  RefAfFile,
  RefPrevalence,
  formula,
  SNPnum = 10000,
  sparseGRM = NULL
)
}
\arguments{
\item{GenoFile}{A character string of the genotype file. See Details section for more details.}

\item{GenoFileIndex}{Additional index file(s) corresponding to GenoFile. See Details section for more details.}

\item{Geno.mtx}{a numeric genotype matrix with each row as an individual and each column as a genetic variant.}

\item{control}{A list of parameters to decide which markers to extract. See \code{Details} section for more details.}

\item{PhenoFile}{A character string of the phenotype file. The phenotype file must have at least three columns: the column of personal identifiers for all individuals, the column of whether the event occurred (0 or 1 or NA), the column of the time of occurrence.}

\item{formula}{a formula object, with the response on the left of \code{Surv(SurvTime , Indicator) ~ covariates} and the covariates on the right. e.g., \code{Surv(SurvTime , Indicator) ~ Cov1 + Cov2}}

\item{SNPnum}{An integer specifying the minimum number of markers. The default is 1e4.}

\item{sparseGRM}{a three-column sparse GRM file with the first column as "ID1",the second column as "ID2", and the last column as "Value" (i.e., two times of kinship coefficient) without information of distant genetic relatedness (such as population structure).}

\item{RefAFfile}{A character string of the reference file. The reference file must be a \code{txt} file (header required) including at least 7 columns: \code{CHROM}, \code{POS}, \code{ID}, \code{REF}, \code{ALT}, \code{AF_ref}, \code{AN_ref}.}

\item{RefPrev}{A numeric value of the event rate in the population.}
}
\value{
An R object with a class of "TestforBatchEffect".
\itemize{
\item{mergeGenoInfo}: A dataframe of marker info and reference MAF.
\item{cutoff}: A numeric, the cut-off of batch effect.
\item{count}: A dataframe of the frequency of the batch effect p-value.
\item{PhenoData}: A dataframe of the input PhenoData.
\item{control}: A list of parameters to decide which markers to extract.
}
}
\description{
This function performs quality control to test for the batch effect between a study cohort and a reference population. And fit a weighted null model.
}
\examples{
setwd(system.file("", package = "WtCoxG"))
RefPrevalence = 0.1

obj.WtCoxG = TestforBatchEffect(GenoFile = "simuBGEN1.bgen",
                             GenoFileIndex = c("simuBGEN1.bgen.bgi",
                                                "simuBGEN1.sample"),
                             OutputFile = "qcBGEN1.txt",
                             control=list(AlleleOrder = "ref-first",
                                          AllMarkers = T,
                                          IndicatorColumn = "SurvEvent", SampleIDColumn = "IID", SurvTimeColumn = "SurvTime"), # specify the column names of sampleID, event, and time
                             PhenoFile = "simuPHENO_WtSPAG.txt",
                             RefAfFile = "RefMAFs1.txt",
                             RefPrevalence = RefPrevalence,
                             formula = Surv(SurvTime , Indicator) ~ Cov1 + Cov2,
                             SNPnum=1e4)
names(obj.WtCoxG)
}
