% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WtCoxG.R
\name{WtCoxG}
\alias{WtCoxG}
\title{WtCoxG method in WtCoxG package}
\usage{
WtCoxG(
  GenoFile,
  GenoFileIndex = NULL,
  Geno.mtx = NULL,
  obj.WtCoxG = NULL,
  PhenoFile = NULL,
  mergeGenoInfoFile = NULL,
  RefPrevalence = NULL,
  control = NULL,
  OutputFile,
  cutoff = 0.1
)
}
\arguments{
\item{GenoFile}{a character of genotype file, .bed file or .bgen file. See Details section for more details.}

\item{GenoFileIndex}{additional index file(s) corresponding to GenoFile. See Details section for more details.}

\item{Geno.mtx}{a numeric genotype matrix with each row as an individual and each column as a genetic variant.}

\item{obj.WtCoxG}{a object with a class of "QCforBatchEffect", which is a list of \code{PhenoData}, \code{mergeGenoInfo} and \code{RefPrevalence}.}

\item{PhenoFile}{A character string of the phenotype file. In addition to inputting \code{obj.WtCoxG}, users can also provide the \code{PhenoFile}, \code{mergeGenoInfoFile}, and \code{RefPrevalence} separately. The phenotype file must have at least three columns: the column of personal identifiers for all individuals, the column of whether the event occurred (0 or 1 or NA), the column of the time of occurrence.}

\item{mergeGenoInfoFile}{A character string of the external MAF file, which must be the results of \code{\link{WtCoxG::QCforBatchEffect}}}

\item{RefPrevalence}{A numeric value of the event rate in the population.}

\item{control}{A list of parameters to decide which markers to extract. See \code{Details} section for more details.}

\item{cutoff}{a numeric to decide the threshold of batch effect p-value. WtCoxG performs GWAS analysis on variants with batch effect p value exceeding the cutoff. The default is 0.1.}
}
\value{
an dataframe. including \code{WtCoxG.ext}, which utilizes external MAF, and \code{WtCoxG.noext} without external MAFs
}
\description{
WtCoxG method is an association testing approach to analyzing complex traits (including but not limited to time-to-event trait) by leveraging external MAFs.
}
\details{
\subsection{The details of \code{control} can be seen in \code{?GRAB.ReadGeno}}{
}
}
\examples{
setwd(system.file("", package = "WtCoxG"))
PhenoData = read.table("simuPHENO_WtSPAG.txt", header = T)
RefPrevalence = 0.1
#step0&1: fit a null model and estimate parameters according to batch effect p values
obj.WtCoxG = QCforBatchEffect(GenoFile = "simuBGEN1.bgen",
                             GenoFileIndex = c("simuBGEN1.bgen.bgi",
                                                "simuBGEN1.sample"),
                             OutputFile = "qcBGEN1.txt",
                             control=list(AlleleOrder = "ref-first",
                                          AllMarkers = T,
                                          IndicatorColumn = "SurvEvent", SampleIDColumn = "IID"),
                             PhenoFile = "simuPHENO_WtSPAG.txt",
                             RefAfFile = "RefMAFs.txt",
                             RefPrevalence = RefPrevalence,
                             formula = Surv(SurvTime , Indicator) ~ Cov1 + Cov2,
                             SNPnum=1e4)
names(obj.WtCoxG)
#step2: conduct association testing
GWAS = WtCoxG(GenoFile = "simuBGEN1.bgen",
            GenoFileIndex = c("simuBGEN1.bgen.bgi", "simuBGEN1.sample"),
            obj.WtCoxG = obj.WtCoxG,
            OutputFile = "simuBGEN1.txt",
            control = list(AlleleOrder = "ref-first", AllMarkers=T))

# Or users can input PhenoFile, mergeGenoInfoFile and RefPrevalence seperately
GWAS = WtCoxG(GenoFile = "simuBGEN1.bgen",
              GenoFileIndex = c("simuBGEN1.bgen.bgi", "simuBGEN1.sample"),
              #obj.WtCoxG = obj.WtCoxG,
              PhenoFile = "simuPHENO_Resid.txt",
              mergeGenoInfoFile = "qcBGEN1.txt",
              RefPrevalence = 0.1,
              OutputFile = "simuBGEN1.txt",
              control = list(AlleleOrder = "ref-first", AllMarkers=T))
}
