% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WtCoxG.R
\name{WtCoxG}
\alias{WtCoxG}
\title{WtCoxG method in WtCoxG package}
\usage{
WtCoxG(
  GenoFile,
  GenoFileIndex = NULL,
  Geno.mtx = NULL,
  obj.WtCoxG,
  OutputFile,
  control = list(AlleleOrder = "ref-first", AllMarkers = T),
  cutoff = 0.1
)
}
\arguments{
\item{GenoFile}{a character of genotype file, .bed file or .bgen file. See Details section for more details.}

\item{GenoFileIndex}{additional index file(s) corresponding to GenoFile. See Details section for more details.}

\item{Geno.mtx}{a numeric genotype matrix with each row as an individual and each column as a genetic variant.}

\item{obj.WtCoxG}{a object with a class of "QCforBatchEffect"}

\item{control}{a list of parameters to decide which markers to extract.See \code{Details} section for more details}

\item{cutoff}{a numeric to decide the threshold of batch effect p-value. WtCoxG performs GWAS analysis on variants with batch effect p value exceeding the cutoff. The default is 0.1.}
}
\value{
an dataframe. including \code{WtCoxG.ext}, which utilizes external MAF, and \code{WtCoxG.noext} without external MAFs
}
\description{
WtCoxG method is an association testing approach to analyzing complex traits (including but not limited to time-to-event trait) by leveraging external MAFs.
}
\details{
\subsection{The details of \code{control} can be seen in \code{?GRAB.ReadGeno}}{
}
}
\examples{
setwd(system.file("", package = "WtCoxG"))
PhenoData = read.table("simuPHENO_WtSPAG.txt", header = T)
RefPrevalence = 0.1
#step0&1: fit a null model and estimate parameters according to batch effect p values
obj.WtCoxG = QCforBatchEffect(GenoFile = "simuBGEN1.bgen",
                             GenoFileIndex = c("simuBGEN1.bgen.bgi",
                                                "simuBGEN1.sample"),
                             OutputFile = "qcBGEN1.txt",
                             control=list(AlleleOrder = "ref-first",
                                          AllMarkers = T,
                                          IndicatorColumn = "SurvEvent", SampleIDColumn = "IID"),
                             PhenoData=PhenoData,
                             RefAfFile = "RefMAFs.txt",
                             RefPrevalence = RefPrevalence,
                             SNPnum=1e4)
names(obj.WtCoxG)
#step2: conduct association testing
GWAS = WtCoxG(GenoFile = "simuBGEN1.bgen",
            GenoFileIndex = c("simuBGEN1.bgen.bgi", "simuBGEN1.sample"),
            obj.WtCoxG = obj.WtCoxG,
            OutputFile = "simuBGEN1.txt",
            control = list(AlleleOrder = "ref-first", AllMarkers=T))
}
